check command proposal
