+++ cli/src/__tests__/commands/error-handling.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, jest, beforeEach } from "@jest/globals";
 import { Command } from "commander";
 import { initCommand } from "../../commands/init";
 import { previewCommand } from "../../commands/preview";
-import { checkCommand } from "../../commands/check";
+import { checkCommand } from "../../commands/check/check";
 
 // Mock console.error and process.exit
 const mockConsoleError = jest.spyOn(console, "error").mockImplementation(() => {});
