+++ cli/src/index.ts
@@ -3,7 +3,7 @@
 import { program } from "commander";
 import { initCommand } from "./commands/init";
 import { previewCommand } from "./commands/preview";
-import { checkCommand } from "./commands/check";
+import { checkCommand } from "./commands/check/check";
 import { compileCommand } from "./commands/compile";
 
 program.name("cg").description("CommonGrants CLI tools").version("0.1.0");
