+++ cli/src/types/command-args.ts
@@ -16,14 +16,6 @@ export const PreviewArgsSchema = z.object({
   }),
 });
 
-export const CheckApiArgsSchema = z.object({
-  apiUrl: z.string().url(),
-  specPath: z.string().endsWith(".tsp").or(z.string().endsWith(".yaml")),
-});
-
-export const CheckSpecArgsSchema = z.object({
-  specPath: z.string().endsWith(".tsp").or(z.string().endsWith(".yaml")),
-});
 
 export const GenerateArgsSchema = z.object({
   specPath: z.string().endsWith(".tsp").or(z.string().endsWith(".yaml")),
@@ -47,11 +39,7 @@ export const AddFieldCommandSchema = z.object({
   description: z.string().optional(),
 });
 
-export const CheckApiCommandSchema = z.object({
-  client: z.string().optional(),
-  report: z.enum(["json", "html"]).optional(),
-  auth: z.string().optional(),
-});
+
 
 export const GenerateServerCommandSchema = z.object({
   lang: z.string().optional(),
@@ -64,9 +52,6 @@ export const GenerateServerCommandSchema = z.object({
     }),
 });
 
-export const CheckSpecCommandSchema = z.object({
-  base: z.string().optional(),
-});
 
 export const GenerateClientCommandSchema = z.object({
   lang: z.string().optional(),
@@ -80,8 +65,6 @@ export const GenerateClientCommandSchema = z.object({
 
 export type AddFieldArgs = z.infer<typeof AddFieldArgsSchema>;
 export type PreviewArgs = z.infer<typeof PreviewArgsSchema>;
-export type CheckApiArgs = z.infer<typeof CheckApiArgsSchema>;
-export type CheckSpecArgs = z.infer<typeof CheckSpecArgsSchema>;
 export type GenerateArgs = z.infer<typeof GenerateArgsSchema>;
 export type CompileArgs = z.infer<typeof CompileArgsSchema>;
 
@@ -91,7 +74,5 @@ export type CompileArgs = z.infer<typeof CompileArgsSchema>;
 
 export type InitCommandOptions = z.infer<typeof InitCommandSchema>;
 export type AddFieldCommandOptions = z.infer<typeof AddFieldCommandSchema>;
-export type CheckApiCommandOptions = z.infer<typeof CheckApiCommandSchema>;
 export type GenerateServerCommandOptions = z.infer<typeof GenerateServerCommandSchema>;
-export type CheckSpecCommandOptions = z.infer<typeof CheckSpecCommandSchema>;
 export type GenerateClientCommandOptions = z.infer<typeof GenerateClientCommandSchema>;
